# swagger.yaml
openapi: 3.0.3
info:
  title: GSM Arena Scraper API
  description: |
    A comprehensive API for scraping mobile device data from GSM Arena.
    This API provides endpoints to scrape brands, models, and detailed specifications
    of mobile devices including RAM, storage, colors, and technical specifications.
  version: 1.0.0
  contact:
    name: GSM Arena Scraper API
    url: https://github.com/your-repo/gsm-arena-scraper
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3002
    description: Development server
  - url: https://your-domain.com
    description: Production server

paths:
  /health:
    get:
      summary: Health Check
      description: Check if the API is running and healthy
      tags:
        - System
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "GSM Arena Scraper API is running"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2025-10-25T15:30:00.000Z"
                  version:
                    type: string
                    example: "1.0.0"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /status:
    get:
      summary: Get Scraping Status
      description: Get the current status of the scraping service
      tags:
        - System
      responses:
        '200':
          description: Status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ready"
                  isRunning:
                    type: boolean
                    example: false
                  hasBrowser:
                    type: boolean
                    example: false
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /brands:
    get:
      summary: Get Available Brands
      description: Get list of all available brands that can be scraped
      tags:
        - Brands
      responses:
        '200':
          description: Brands retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Brand'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /brands/all:
    get:
      summary: Get All Brands (Detailed)
      description: Get detailed information about all available brands
      tags:
        - Brands
      responses:
        '200':
          description: All brands retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Brand'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /brands/scrape:
    post:
      summary: Scrape Brands
      description: |
        Scrape data for specified brands or all brands if none specified.
        This is the main endpoint for scraping mobile device data.
      tags:
        - Scraping
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                brands:
                  type: array
                  items:
                    type: string
                  description: Array of brand names to scrape (optional - if empty, scrapes all brands)
                  example: ["apple", "samsung", "xiaomi"]
                options:
                  $ref: '#/components/schemas/ScrapingOptions'
      responses:
        '200':
          description: Scraping completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Brand scraping completed successfully"
                  data:
                    $ref: '#/components/schemas/ScrapingResult'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /brands/{brandName}/scrape:
    post:
      summary: Scrape Specific Brand Models
      description: Scrape models for a specific brand
      tags:
        - Scraping
      parameters:
        - name: brandName
          in: path
          required: true
          description: Name of the brand to scrape
          schema:
            type: string
            example: "apple"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                options:
                  $ref: '#/components/schemas/ScrapingOptions'
      responses:
        '200':
          description: Brand scraping completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/BrandResult'
        '400':
          description: Bad request - Brand name is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /brands/{brandName}/devices:
    get:
      summary: Get Devices by Brand
      description: Get list of devices for a specific brand
      tags:
        - Devices
      parameters:
        - name: brandName
          in: path
          required: true
          description: Name of the brand
          schema:
            type: string
            example: "apple"
      responses:
        '200':
          description: Devices retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /devices/{deviceId}/specifications:
    get:
      summary: Get Device Specifications
      description: Get detailed specifications for a specific device
      tags:
        - Devices
      parameters:
        - name: deviceId
          in: path
          required: true
          description: ID of the device
          schema:
            type: string
            example: "12345"
      responses:
        '200':
          description: Device specifications retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/DeviceSpecifications'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /devices/search:
    get:
      summary: Search Devices by Keyword
      description: Search for devices using a keyword
      tags:
        - Devices
      parameters:
        - name: keyword
          in: query
          required: true
          description: Search keyword
          schema:
            type: string
            example: "iPhone 15"
      responses:
        '200':
          description: Search results retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Device'
        '400':
          description: Bad request - Keyword parameter is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /scrape/test:
    post:
      summary: Test Scraping
      description: Test the scraping functionality with a single brand
      tags:
        - Scraping
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                brandName:
                  type: string
                  description: Brand name to test (defaults to 'apple')
                  example: "apple"
      responses:
        '200':
          description: Test scraping completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/TestResult'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /data/latest:
    get:
      summary: Get Latest Data
      description: Get the latest scraped data
      tags:
        - Data
      responses:
        '200':
          description: Latest data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Latest data retrieved"
                  data:
                    type: object
                    nullable: true
                    example: null
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /data/save:
    post:
      summary: Save Data
      description: Save scraped data to storage
      tags:
        - Data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - data
              properties:
                data:
                  type: object
                  description: Data to save
                filename:
                  type: string
                  description: Optional filename for saving
                  example: "gsm-data.json"
      responses:
        '200':
          description: Data saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Data saved successfully"
                  filename:
                    type: string
                    example: "gsm-data.json"
        '400':
          description: Bad request - Data is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Brand:
      type: object
      properties:
        name:
          type: string
          example: "apple"
        url:
          type: string
          example: "https://www.gsmarena.com/apple-phones-48.php"
        logo_url:
          type: string
          example: "https://www.gsmarena.com/img/logo_apple.png"
        is_active:
          type: boolean
          example: true

    Device:
      type: object
      properties:
        name:
          type: string
          example: "iPhone 17 Pro Max"
        url:
          type: string
          example: "https://www.gsmarena.com/iphone_17_pro_max-12345.php"
        year:
          type: integer
          example: 2025

    DeviceSpecifications:
      type: object
      properties:
        specifications:
          type: object
          additionalProperties:
            type: string
          example:
            "Announced": "2025, September 09"
            "Display Type": "LTPO Super Retina XDR OLED, 120Hz"
            "Internal": "256GB 12GB RAM"
        image_url:
          type: string
          example: "https://www.gsmarena.com/img/phones/iphone_17_pro_max.jpg"
        ram_options:
          type: array
          items:
            type: string
          example: ["12"]
        storage_options:
          type: array
          items:
            type: string
          example: ["256GB", "512GB", "1TB", "2TB"]
        color_options:
          type: array
          items:
            type: string
          example: ["Silver", "Cosmic Orange", "Deep Blue"]

    Model:
      type: object
      properties:
        brand_name:
          type: string
          example: "apple"
        model_name:
          type: string
          example: "iPhone 17 Pro Max"
        series:
          type: string
          example: "iPhone"
        ram_options:
          type: array
          items:
            type: string
          example: ["12"]
        storage_options:
          type: array
          items:
            type: string
          example: ["256GB", "512GB", "1TB", "2TB"]
        color_options:
          type: array
          items:
            type: string
          example: ["Silver", "Cosmic Orange", "Deep Blue"]
        specifications:
          type: object
          additionalProperties:
            type: string
        release_date:
          type: string
          format: date
          example: "2025-01-01"
        image_url:
          type: string
          example: "https://www.gsmarena.com/img/phones/iphone_17_pro_max.jpg"
        is_active:
          type: boolean
          example: true

    BrandResult:
      type: object
      properties:
        brand:
          type: object
          properties:
            name:
              type: string
              example: "apple"
            logo_url:
              type: string
              example: "https://www.gsmarena.com/img/logo_apple.png"
            is_active:
              type: boolean
              example: true
            models:
              type: array
              items:
                $ref: '#/components/schemas/Model'
        scraped_at:
          type: string
          format: date-time
          example: "2025-10-25T15:30:00.000Z"
        total_models:
          type: integer
          example: 5
        options:
          $ref: '#/components/schemas/ScrapingOptions'

    ScrapingResult:
      type: object
      properties:
        brands:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "apple"
              logo_url:
                type: string
                example: "https://www.gsmarena.com/img/logo_apple.png"
              is_active:
                type: boolean
                example: true
              models:
                type: array
                items:
                  $ref: '#/components/schemas/Model'
        scraped_at:
          type: string
          format: date-time
          example: "2025-10-25T15:30:00.000Z"
        total_brands:
          type: integer
          example: 1
        total_models:
          type: integer
          example: 5
        options:
          $ref: '#/components/schemas/ScrapingOptions'

    TestResult:
      type: object
      properties:
        brand:
          $ref: '#/components/schemas/BrandResult'
        scraped_at:
          type: string
          format: date-time
          example: "2025-10-25T15:30:00.000Z"
        total_models:
          type: integer
          example: 3
        test_mode:
          type: boolean
          example: true

    ScrapingOptions:
      type: object
      properties:
        minYear:
          type: integer
          description: Minimum year filter for devices
          example: 2023
        modelsPerBrand:
          type: integer
          description: Maximum number of models to scrape per brand
          example: 5
        excludeKeywords:
          type: array
          items:
            type: string
          description: Keywords to exclude from scraping
          example: ["prototype", "concept"]

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
        error:
          type: string
          example: "Detailed error message"
        statusCode:
          type: integer
          example: 500

tags:
  - name: System
    description: System health and status endpoints
  - name: Brands
    description: Brand-related operations
  - name: Devices
    description: Device-related operations
  - name: Scraping
    description: Data scraping operations
  - name: Data
    description: Data management operations